updated code as per 6-2-3 and added the start to login page
My code for some reason likes to put the new post at the bottom until I refresh the page, I checked it against other classmate's code and mine is the same. as a temporary fix I have location.reload at the bottom to make sure the new post goes to the top.